:books: [Docs]: Update LoadingComponent.tsx
- Create directory of LoadingComponent
- Add file of style.ts in LoadingComponent directory
